feat: Replace future steps with simple placeholders and remove unused code (closes #111)
- Replace _migrate_code, _validate_migration, _repair_issues, and _finalize_migration with simple log placeholders
- Remove entire agent system (agent.py, AgentManager, AgentType) as it's no longer used
- Remove agent-related prompt files (analyzer.txt, migrator.txt, repairer.txt, tester.txt)
- Update coordinator.py to remove agent imports and references
- Update cli.py to remove agent cleanup references
- Remove tests for removed functionality (test_orchestration.py)
- Add focused tests for simplified coordinator methods
- Update conftest.py to remove agent-related mocks

This significantly simplifies the codebase by removing complex unused code while maintaining the working test discovery and test execution steps.
